Hotel Prices
This dataset addresses Data Science students and/or Beginners who want to dive into Regression or Clustering without the need to pre-clean the data first.

Content
This dataset consists of a pre-cleaned .csv table that has been translated from German to English.

There are four columns in this dataset:

Profit (How much money does this hotel make in a year)
Price in Millions (€)
Square Meter (Hotel Area)
City
Here, "Hotel Prices" does not refer to the cost of spending a night at those hotels but the price for buying them. This would be an interesting chart for someone who wants to buy a hotel and needs to judge whether he/she is overpaying or getting a great deal depending on similar objects in other comparable cities.